New bindings

+++ src/constants/ETHER_TYPE.rs
@@ -2,12 +2,12 @@
 // Copyright © 2016 The developers of dpdk-sys. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/dpdk-sys/master/COPYRIGHT.
 
 
-pub const ETHER_TYPE_1588: c_int = 35063;
-pub const ETHER_TYPE_ARP: c_int = 2054;
-pub const ETHER_TYPE_IPv4: c_int = 2048;
-pub const ETHER_TYPE_IPv6: c_int = 34525;
-pub const ETHER_TYPE_LEN: c_int = 2;
-pub const ETHER_TYPE_RARP: c_int = 32821;
-pub const ETHER_TYPE_SLOW: c_int = 34825;
-pub const ETHER_TYPE_TEB: c_int = 25944;
-pub const ETHER_TYPE_VLAN: c_int = 33024;
+pub const ETHER_TYPE_1588: uint16_t = 35063;
+pub const ETHER_TYPE_ARP: uint16_t = 2054;
+pub const ETHER_TYPE_IPv4: uint16_t = 2048;
+pub const ETHER_TYPE_IPv6: uint16_t = 34525;
+pub const ETHER_TYPE_LEN: uint16_t = 2;
+pub const ETHER_TYPE_RARP: uint16_t = 32821;
+pub const ETHER_TYPE_SLOW: uint16_t = 34825;
+pub const ETHER_TYPE_TEB: uint16_t = 25944;
+pub const ETHER_TYPE_VLAN: uint16_t = 33024;
